Refactor area-line-chart to stacked-bar-chart
* Very weird errors with the line chart (incorrect data visualisation)
* With a maximum amount of datapoints (due to decimisation) the bar-chart looks acceptable with a lot of values
