+++ frontend/src/app/pages/exercises/exercise/shared/exercise-statistics/exercise-statistics.module.ts
@@ -3,12 +3,12 @@ import { CommonModule } from '@angular/common';
 import { NgbDropdownModule } from '@ng-bootstrap/ng-bootstrap';
 import { SharedModule } from 'src/app/shared/shared.module';
 import { ExerciseStatisticsModalComponent } from './exercise-statistics-modal/exercise-statistics-modal.component';
-import { TimeLineAreaChartComponent } from './time-line-area-chart/time-line-area-chart.component';
+import { StackedBarChartComponent } from './stacked-bar-chart/stacked-bar-chart.component';
 
 @NgModule({
     declarations: [
         ExerciseStatisticsModalComponent,
-        TimeLineAreaChartComponent,
+        StackedBarChartComponent,
     ],
     imports: [CommonModule, NgbDropdownModule, SharedModule],
 })
