+++ frontend/src/app/pages/exercises/exercise/shared/exercise-statistics/stacked-bar-chart/stacked-bar-chart.component.html
