+++ frontend/src/app/pages/exercises/exercise/shared/exercise-statistics/stacked-bar-chart/stacked-bar-chart.component.ts
@@ -1,20 +1,20 @@
 import type { AfterViewInit, OnChanges, OnDestroy } from '@angular/core';
 import { Component, ElementRef, Input, NgZone, ViewChild } from '@angular/core';
 import type { SimpleChangesGeneric } from 'src/app/shared/types/simple-changes-generic';
-import type { TimeLineAreaChartDatasets } from './time-line-area-chart';
-import { TimeLineAreaChart } from './time-line-area-chart';
+import type { StackedBarChartDatasets } from './time-line-area-chart';
+import { StackedBarChart } from './time-line-area-chart';
 
 @Component({
-    selector: 'app-time-line-area-chart',
-    templateUrl: './time-line-area-chart.component.html',
-    styleUrls: ['./time-line-area-chart.component.scss'],
+    selector: 'app-stacked-bar-chart',
+    templateUrl: './stacked-bar-chart.component.html',
+    styleUrls: ['./stacked-bar-chart.component.scss'],
 })
-export class TimeLineAreaChartComponent
+export class StackedBarChartComponent
     implements AfterViewInit, OnChanges, OnDestroy
 {
     @Input() statistics!: {
         labels: string[];
-        datasets: TimeLineAreaChartDatasets;
+        datasets: StackedBarChartDatasets;
     };
 
     @ViewChild('chart', { static: true })
@@ -22,12 +22,12 @@ export class TimeLineAreaChartComponent
 
     constructor(private readonly ngZone: NgZone) {}
 
-    private chart?: TimeLineAreaChart;
+    private chart?: StackedBarChart;
 
     ngAfterViewInit() {
         // Run outside angular zone for improved performance
         this.ngZone.runOutsideAngular(() => {
-            this.chart = new TimeLineAreaChart(this.chartCanvas.nativeElement);
+            this.chart = new StackedBarChart(this.chartCanvas.nativeElement);
         });
         this.updateChartData();
     }
@@ -52,7 +52,7 @@ export class TimeLineAreaChartComponent
     }
 }
 
-export interface TimeLineAreaChartStatistics {
+export interface StackedBarChartStatistics {
     labels: string[];
-    datasets: TimeLineAreaChartDatasets;
+    datasets: StackedBarChartDatasets;
 }
